Assignment 2: Time to Practice: Working with Events & State
Here we add expenses filter dropdown component named `ExpensesFilter.js`.  You
select the year.  The assignment was to:

- Listen to changes made to the dropdown, so I used the `onChange` event listener
  on the `select` element.  This along with an event handler.

- Next was to take that dropdown selection event value and pass it up the chain
  to the parent `Expenses.js` component using a props function defined in the
  `Expenses.js` component.

- Next was to take the data that was passed to the `Expenses.js` component
  function (which was called via the `ExpensesFilter.js` component *props*) and
  save that data in a stateful value, which I've named `filteredYear`.

- A bonus assignment (which he only told us about in the explanation follow-up
  video) was to set a default value for the dropdown via two-way binding.  This
  was accomplished by passing the stateful value back down the chain from the
  `Expenses.js` component down to the `ExpensesFilter.js` component via a *props*
  value, which is then set on the `select` element as the `value` attribute.

NOTE: The fact that we used two-way binding ***and*** the fact that we are storing
the stateful value in the parent `Expenses.js` component, makes the `ExpensesFilter.js`
component a special type of component known as a **"Controlled Component"**.
The parent essentially **controls** the child component.  The `ExpensesFilter.js`
component is simply a presentation component, of sorts.

Additionally, another concept is **"Stateful Components"** vs. **"Stateless Components".
**"Stateful Components"** are those that manage states.  **"Stateless Components"**
are those that don't manage any states.  The `ExpensesFilter.js` component is a
**"Stateless Component"**, while the `Expenses.js` component is now a **"Stateful Component"**,
since we're managing the Expenses Filter state there.